[fix(Channel)] 채팅방 폭파 소켓메세지 역직렬화 예외 터지던 오류 수정
